+++ iOS Example Obj-C/iOS Example Obj-C/AppDelegate.h
@@ -1,0 +1,17 @@
+//
+//  AppDelegate.h
+//  iOS Example Obj-C
+//
+//  Created by Hamilton Chapman on 09/09/2016.
+//  Copyright © 2016 Pusher. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+@interface AppDelegate : UIResponder <UIApplicationDelegate>
+
+@property (strong, nonatomic) UIWindow *window;
+
+
+@end
+